A5: implementa SQLite setup + migrations
Persistência local via sqlx (pool SqlitePool) com o schema completo
de docs/schema.md. O init_db() roda no setup() do Tauri, cria
~/.genesis/genesis.db se faltar e registra o pool no managed state
pra que as commands futuras o injetem via State<DbPool>.

- Cargo.toml: sqlx 0.8 (sqlite, runtime-tokio) + dirs 5
- migrations/001_init.sql: 4 tabelas, 9 índices (projects,
  executions, steps, messages), 3 triggers (updated_at automático,
  contador de completed_steps, auto-complete de execution). Tudo
  IF NOT EXISTS — idempotente a cada startup
- db/mod.rs: DbPool = SqlitePool, db_path() resolve ~/.genesis/
  genesis.db, init_db() aplica pragmas WAL + foreign_keys +
  busy_timeout(5s) via SqliteConnectOptions e executa as migrations
  num Executor multi-statement
- db/models.rs: Project, Execution, ExecutionStep, ChatMessage com
  derive(FromRow); campos INTEGER viram i64 (SQLite armazena como
  i64 — i32 quebraria o FromRow)
- db/queries.rs: CRUD completo, SQL inline só aqui (CLAUDE.md):
  list/get/insert/delete projects, list/insert/update_status
  executions, list/insert/update_status steps, list/insert messages
- lib.rs: .setup() chama init_db() via async_runtime::block_on e
  app.manage(pool); evita race com commands que leem o pool

Verificado com sqlite3 ~/.genesis/genesis.db: .tables e .schema
batem com docs/schema.md, journal_mode=wal, 3 triggers presentes.

+++ src-tauri/src/db/models.rs
@@ -1,6 +1,12 @@
-use serde::{Deserialize, Serialize};
+//! Rust structs mirroring the SQLite schema (see docs/schema.md §3).
+//!
+//! Integer columns use `i64` because SQLite stores every INTEGER as a 64-bit
+//! signed value; sqlx's `FromRow` rejects `i32` unless you convert explicitly.
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+use serde::{Deserialize, Serialize};
+use sqlx::FromRow;
+
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct Project {
     pub id: String,
     pub name: String,
@@ -9,7 +15,7 @@ pub struct Project {
     pub updated_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct Execution {
     pub id: String,
     pub project_id: String,
@@ -17,30 +23,30 @@ pub struct Execution {
     pub status: String,
     pub started_at: Option<String>,
     pub finished_at: Option<String>,
-    pub total_steps: i32,
-    pub completed_steps: i32,
+    pub total_steps: i64,
+    pub completed_steps: i64,
     pub created_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct ExecutionStep {
     pub id: String,
     pub execution_id: String,
     pub step_id: String,
-    pub step_order: i32,
+    pub step_order: i64,
     pub tool: String,
     pub status: String,
     pub input: String,
     pub output: Option<String>,
     pub error: Option<String>,
-    pub retries: i32,
+    pub retries: i64,
     pub started_at: Option<String>,
     pub finished_at: Option<String>,
     pub duration_ms: Option<i64>,
     pub created_at: String,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct ChatMessage {
     pub id: String,
     pub execution_id: Option<String>,
